requirements.txt
pysteps~=1.6.1
folium~=0.12.1.post1
numpy~=1.22.2
requests~=2.27.1
pillow~=9.0.1
beautifulsoup4~=4.10.0
cartopy~=0.18.0
branca~=0.4.2
matplotlib~=3.5.1
tqdm~=4.62.3
pyproj~=2.6.1.post1
pygrib~=2.1.4
typer~=0.4.0
git-python==1.0.3
scikit-image==0.19.2
opencv-python==4.5.5.62
pandas==1.4.1
plotly==5.6.0
